+++ src/lib/tuya.ts
@@ -1497,7 +1497,7 @@ const tuyaTz = {
 
                 if (dpEntry[3] && dpEntry[3].optimistic === false) continue;
 
-                state[key] = value;
+                state[attr] = value;
             }
             return {state};
         },
@@ -1800,7 +1800,7 @@ export function getHandlersForDP(
         : [
               {
                   key: [name],
-                  endpoint: endpoint,
+                  endpoints: [endpoint],
                   convertSet: async (entity, key, value, meta) => {
                       // A set converter is only called once; therefore we need to loop
                       const state: KeyValue = {};